chore: refactored a bit.
