+++ apps/web/src/app/tracks/_components/track-enrolled-section.tsx
@@ -1,20 +1,17 @@
+import { getServerAuthSession } from '@repo/auth/server';
+import { prisma } from '@repo/db';
 import Link from 'next/link';
-import { PersonalTrackCard } from './personal-track-card';
-import type { EnrolledTracks } from '.';
 import { Carousel } from '~/components/ui/carousel';
+import { PersonalTrackCard } from './personal-track-card';
 
-interface EnrolledTrackSectionProps {
-  title: string;
-  tracks: EnrolledTracks;
-}
-
-export function EnrolledTrackSection({ title, tracks }: EnrolledTrackSectionProps) {
+export async function EnrolledTrackSection() {
+  const tracks = await getUserEnrolledTracks();
   return (
     <div>
       <div className="container flex items-center justify-between gap-3 px-4 pt-5">
         <h2 className="relative text-3xl font-bold tracking-tight">
           <div className="absolute -left-8 -z-10 h-12 w-32 rounded-full bg-blue-300/50 blur-3xl" />
-          {title}
+          Your Tracks
         </h2>
       </div>
       <section className="relative flex w-full flex-row gap-4 overflow-hidden rounded-[2.5rem]">
@@ -40,4 +37,45 @@ export function EnrolledTrackSection({ title, tracks }: EnrolledTrackSectionProp
       </section>
     </div>
   );
+}
+
+export type EnrolledTracks = Awaited<ReturnType<typeof getUserEnrolledTracks>>;
+
+/**
+ * Fetches user enrolled tracks based on current session.
+ */
+async function getUserEnrolledTracks() {
+  const session = await getServerAuthSession();
+  return prisma.track.findMany({
+    where: {
+      enrolledUsers: {
+        some: {
+          id: session?.user.id,
+        },
+      },
+    },
+    include: {
+      trackChallenges: {
+        include: {
+          challenge: {
+            include: {
+              submission: {
+                where: {
+                  userId: session?.user.id,
+                },
+              },
+            },
+          },
+        },
+      },
+      _count: {
+        select: {
+          enrolledUsers: true,
+        },
+      },
+    },
+    orderBy: {
+      title: 'asc',
+    },
+  });
 }
